requirements.txt
pytest
pytest_bdd
requests
Faker
scoop==0.7.2.0
webdriver-manager
tox
allure-pytest
allure-pytest-bdd
pytest-env
mysql
playwright==1.46.0
pytest-playwright
axe-playwright-python
axe-core-python
pytest-xdist
asyncio
pytest-custom-exit-code
allure-combine
pytest-retry
python-dateutil
